Print next steps on exit
These new messages duplicate the instructions in the README. This is
intentional, as many people will likely run this tool without reading
the instructions first.

The main motivation for this change is an extra reminder to check for
valid /etc/master.passwd, /etc/group, and /etc/ssh/sshd_config after
a error-free conversion. If this reminder prevents at least one broken
system, it is worth it.

Sponsored by:	The FreeBSD Foundation